+++ src/main/java/blackjack/domain/card/Cards.java
@@ -1,8 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
 
 public class Cards {
     private final List<Card> cards;
