+++ src/test/java/blackjack/domain/human/PlayersTest.java
@@ -2,7 +2,6 @@ package blackjack.domain.human;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import blackjack.domain.Name;
 import java.util.List;
 import org.junit.jupiter.api.Test;
